Extract coin lookup fns
