+++ fuel-core/src/main.rs
@@ -5,6 +5,7 @@ use tracing::trace;
 
 mod args;
 pub(crate) mod chain_config;
+pub(crate) mod coin_lookup;
 pub mod database;
 pub(crate) mod executor;
 pub mod model;
